laravel 订单的数据填充，增加address

+++ database/migrations/2021_12_08_232623_create_orders_table.php
@@ -31,7 +31,7 @@ class CreateOrdersTable extends Migration
             $table->string('order_no')->comment('订单单号');
             $table->integer('amount')->comment('总金额，单位：分');
             $table->tinyInteger('status')->default(1)->comment('订单状态 1下单 2支付 3发货 4收货');
-            $table->integer('address_id')->comment('收货地址');
+            $table->bigInteger('address_id')->comment('收货地址');
             $table->string('express_type')->nullable()->comment('快递类型 SF YT YD');
             $table->string('express_no')->nullable()->comment('快递单号');
             $table->timestamp('pay_time')->nullable()->comment('支付时间');
